+++ README.md
@@ -35,7 +35,6 @@ graph TD
     G --> I[재고 반영]
     I --> J[창고 재고 업데이트]
     H --> K[입고 취소]
-
 ```
 
 **입고 상태별 설명:**
