+++ api/src/main/java/inventory/inbound/domain/Inbound.java
@@ -1,6 +1,6 @@
 package inventory.inbound.domain;
 
-import inventory.inbound.enums.InboundStatus;
+import inventory.inbound.domain.enums.InboundStatus;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
