+++ api/src/main/java/inventory/inbound/domain/enums/InboundStatus.java
@@ -1,4 +1,4 @@
-package inventory.inbound.enums;
+package inventory.inbound.domain.enums;
 
 import lombok.Getter;
