+++ api/src/main/java/inventory/inbound/service/request/CreateInboundRequest.java
@@ -1,4 +1,4 @@
-package inventory.inbound.controller.request;
+package inventory.inbound.service.request;
 
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotEmpty;
