+++ api/src/main/java/inventory/product/service/request/CreateProductRequest.java
@@ -1,4 +1,4 @@
-package inventory.product.controller.request;
+package inventory.product.service.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
