+++ api/src/main/java/inventory/product/service/request/UpdateProductRequest.java
@@ -1,4 +1,4 @@
-package inventory.product.controller.request;
+package inventory.product.service.request;
 
 import jakarta.validation.constraints.NotBlank;
