+++ api/src/main/java/inventory/supplier/service/response/SupplierResponse.java
@@ -1,4 +1,4 @@
-package inventory.supplier.controller.response;
+package inventory.supplier.service.response;
 
 import inventory.supplier.domain.Supplier;
