+++ api/src/main/java/inventory/warehouse/service/request/CreateWarehouseRequest.java
@@ -1,4 +1,4 @@
-package inventory.warehouse.controller.request;
+package inventory.warehouse.service.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
