+++ api/src/main/java/inventory/warehouse/service/response/WarehouseResponse.java
@@ -1,4 +1,4 @@
-package inventory.warehouse.controller.response;
+package inventory.warehouse.service.response;
 
 import inventory.warehouse.domain.Warehouse;
