+++ api/src/main/java/inventory/warehouse/service/response/WarehouseStockResponse.java
@@ -1,4 +1,4 @@
-package inventory.warehouse.controller.response;
+package inventory.warehouse.service.response;
 
 import inventory.product.domain.Product;
 import inventory.warehouse.domain.Warehouse;
