+++ api/src/test/java/inventory/inbound/controller/InboundControllerTest.java
@@ -2,14 +2,13 @@ package inventory.inbound.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import inventory.common.exception.GlobalExceptionHandler;
-import inventory.inbound.controller.request.CreateInboundRequest;
-import inventory.inbound.controller.request.InboundProductRequest;
-import inventory.inbound.controller.request.UpdateInboundStatusRequest;
-import inventory.inbound.controller.response.InboundResponse;
-import inventory.inbound.controller.response.InboundProductResponse;
+import inventory.inbound.service.request.CreateInboundRequest;
+import inventory.inbound.service.request.InboundProductRequest;
+import inventory.inbound.service.request.UpdateInboundStatusRequest;
+import inventory.inbound.service.response.InboundResponse;
+import inventory.inbound.service.response.InboundProductResponse;
 import inventory.inbound.domain.Inbound;
-import inventory.inbound.domain.InboundProduct;
-import inventory.inbound.enums.InboundStatus;
+import inventory.inbound.domain.enums.InboundStatus;
 import inventory.inbound.service.InboundService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
